fix map input_type to task_type for vertex ai
